requirements.txt
# Framework
pytest
selenium
Appium-Python-Client
pycmd

# Reporting
allure-pytest

# Fake Data Generator
Faker